보스 사신 재수정

+++ Assets/Scripts/InGame/Arbait/BlueHair.cs
@@ -11,7 +11,7 @@ public class BlueHair : ArbaitBatch {
 	{
 		base.Awake ();
 
-		nIndex = E_ARBAIT.E_BLUEHAIR;
+		nIndex = (int)E_ARBAIT.E_BLUEHAIR;
 
 
 	}
